Sun, Jul 7, 2024, 12:54 PM +00:00

+++ 1.Conditions/interpreter/interpreter.py
@@ -3,8 +3,8 @@ def main():
     expression = input("Expression: ")
 
     x, y, z = expression.split(" ")
-    x = float(x)
-    z = float(z)
+    x = int(x)
+    z = int(z)
 
     result = interpreter(x, y, z)
